Create Computer Speak prank
A simple prank.  It will turn the volume up if down and then repeat the words in "".  These words can be changed to your own liking.  Have fun.File has to be loaded onto target PC. I hide the file then I like to point Google Chrome icon to this file so when the suspect clicks chrome it runs this file.